tracking: moved shared modules from server to lib
