+++ packages/slice-machine/server/src/api/custom-types/index.ts
@@ -6,7 +6,7 @@ import {
   CustomTypeSM,
 } from "@slicemachine/core/build/models/CustomType/index";
 import { CustomType } from "@prismicio/types-internal/lib/customtypes/CustomType";
-import * as IO from "../io";
+import * as IO from "../../../../lib/io";
 import { ClientError } from "@slicemachine/client";
 import { getLocalCustomTypes } from "../../../../lib/utils/customTypes";
