--- packages/slice-machine/server/src/api/env.ts
@@ -1,5 +1,0 @@
-import getEnv from "./services/getEnv";
-
-export default async function handler() {
-  return getEnv();
-}
