+++ packages/slice-machine/server/tsconfig.json
